Refactoring, cleanup
